+++ CupTaster/Cupping/Cupping.swift
@@ -10,7 +10,7 @@ import CoreData
 
 struct CuppingView: View {
     @Environment(\.presentationMode) var presentationMode
-    @Environment(\.managedObjectContext) private var moc
+    @Environment(\.managedObjectContext) var moc
     @Namespace var namespace
     
     @FetchRequest(entity: CuppingForm.entity(), sortDescriptors: []) var cuppingForms: FetchedResults<CuppingForm>
@@ -33,110 +33,9 @@ struct CuppingView: View {
     
     var body: some View {
         ZStack {
-            if cuppingModel.sampleViewVisible {
-                SampleSelectorView(cuppingModel: cuppingModel, namespace: namespace)
-                    .zIndex(1)
-                    .clipped()
-            } else if cuppingModel.samplesEditorActive {
-                Form {
-                    Section {
-                        TextField("Cupping name", text: $cuppingModel.cupping.name)
-                    }
-										
-                    if showCuppingsDatePicker {
-                        Section {
-                            DatePicker("Date", selection: $cuppingModel.cupping.date, displayedComponents: [.date])
-                        }
-                    }
-                    
-                    Section {
-                        ForEach(samples) { sample in
-                            SampleFormRowView(sample: sample)
-                        }
-                        .onMove { indexSet, offset in
-                            var revisedItems: [Sample] = cuppingModel.sortedSamples
-                            revisedItems.move(fromOffsets: indexSet, toOffset: offset)
-                            
-                            for reverseIndex in stride(from: revisedItems.count - 1, through: 0, by: -1) {
-                                revisedItems[reverseIndex].ordinalNumber = Int16(reverseIndex)
-                            }
-                        }
-                        .onDelete { offsets in
-                            for index in offsets {
-                                moc.delete(samples[index])
-                            }
-                        }
-                    }
-                    
-                    Section {
-                        Button {
-                            let usedNames: [String] = cuppingModel.cupping.samples.map { $0.name }
-                            let defaultName: String = SampleNameGenerator().generateSampleDefaultName(usedNames: usedNames)
-                            
-                            let sample: Sample = Sample(context: moc)
-                            
-                            sample.name = defaultName
-                            sample.ordinalNumber = Int16(cuppingModel.cupping.samples.count)
-                            
-                            if let cuppingForm = cuppingModel.cupping.form {
-                                for groupConfig in cuppingForm.qcGroupConfigurations {
-                                    let qcGroup: QCGroup = QCGroup(context: moc)
-                                    qcGroup.sample = sample
-                                    qcGroup.configuration = groupConfig
-                                    for qcConfig in groupConfig.qcConfigurations {
-                                        let qualityCriteria = QualityCriteria(context: moc)
-                                        qualityCriteria.title = qcConfig.title
-                                        qualityCriteria.value = qcConfig.value
-                                        qualityCriteria.group = qcGroup
-                                        qualityCriteria.configuration = qcConfig
-                                    }
-                                }
-                            }
-                            
-                            cuppingModel.cupping.addToSamples(sample)
-                        } label: {
-                            Label("Add sample", systemImage: "plus")
-                        }
-                    }
-                }
-                .environment(\.editMode, .constant(.active))
-                .padding(.bottom, 44) // toolbar
-                .resignKeyboardOnDragGesture() { try? moc.save() }
-                .clipped()
-            } else {
-                ScrollView {
-                    Text(cuppingModel.cupping.name)
-                        .font(.largeTitle)
-                        .fontWeight(.heavy)
-                        .frame(maxWidth: .infinity, alignment: .leading)
-                        .padding([.top, .horizontal], 20)
-                    
-                    LazyVGrid(columns: Array(repeating: .init(.flexible()), count: 2)) {
-                        ForEach(samples) { sample in
-                            Button {
-                                cuppingModel.selectedSample = sample
-                                cuppingModel.selectedSampleIndex = cuppingModel.sortedSamples.firstIndex(of: sample)!
-                                cuppingModel.samplesAppearance = .criteria
-                                cuppingModel.offset = .zero
-                                cuppingModel.switchingToPreviews = false
-                                
-                                DispatchQueue.main.asyncAfter(deadline: .now() + 0.01) {
-                                    withAnimation {
-                                        cuppingModel.sampleViewVisible = true
-                                    }
-                                }
-                            } label: {
-                                SampleView(cuppingModel: cuppingModel, sample: sample).preview
-                            }
-                            .matchedGeometryEffect(id: sample.id, in: namespace)
-                            .zIndex(cuppingModel.selectedSample?.id == sample.id ? 1 : 0)
-                        }
-                    }
-                    .padding([.bottom, .horizontal])
-                    .padding(.bottom, 44) // toolbar
-                }
-                .clipped()
-            }
+            if cuppingModel.sampleViewVisible { SampleSelectorView(cuppingModel: cuppingModel, namespace: namespace) }
+			else if cuppingModel.samplesEditorActive { samplesEditor }
+			else { samplesPreview }
             
             CuppingToolbarView(
                 presentationMode: _presentationMode,
@@ -154,33 +53,9 @@ struct CuppingView: View {
             .ignoresSafeArea(.keyboard, edges: sampleNameTextfieldFocus == nil ? [.all] : [])
             .animation(.default, value: sampleNameTextfieldFocus)
             .zIndex(2)
-            
-            if let qcGroupConfig = cuppingModel.selectedHintsQCGConfig {
-                ScrollView(showsIndicators: false) {
-                    HintsMenuView(qcGroupConfig: qcGroupConfig)
-                        .padding(50)
-                }
-                .background(.ultraThinMaterial)
-                .safeAreaInset(edge: .bottom) {
-                    Button {
-                        withAnimation {
-                            cuppingModel.selectedHintsQCGConfig = nil
-                        }
-                    } label: {
-                        Text("Done")
-                            .foregroundColor(.white)
-                            .padding()
-                            .frame(maxWidth: .infinity)
-                            .background(Color.accentColor)
-                            .cornerRadius(15)
-                            .shadow(radius: 15)
-                            .padding([.horizontal, .bottom], 50)
-                    }
-                }
-                .transition(.move(edge: .bottom))
-                .zIndex(3)
-            }
-            
+			
+			selectedHint
+			
             LinearGradient(
                 colors: [Color(uiColor: .systemGroupedBackground), Color(uiColor: .systemGroupedBackground).opacity(0)],
                 startPoint: .top,
